frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Fahad Ahmed", email = "fahad.ahmed95@live.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.0.0",
    "requests>=2.32.3",
    "streamlit>=1.40.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
